infrastructure/user-management/requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
flask-apispec==0.8.6
pyjwt==1.4.2
SQLAlchemy==1.3.13
psycopg2==2.8.4
marshmallow==3.5.0
SQLAlchemy-Utils==0.36.1
python-slugify==4.0.0
passlib==1.7.2
cryptography==2.9
python-dateutil==2.8.1
marshmallow-enum==1.5.1
